Implement a better work_mem estimation.
Some rules were not implemented, as the 1024 minimal main array size, so
clean up everything and implement a better estimation.  Add more
comments and define some magic constants (or use the one available when
possible) for better readibility.

As a consequence, the regression tests need to be updated since the
estimate is now way closer to reality.